@@ -4,22 +4,16 @@ import { BrowserModule } from '@angular/platform-browser';
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
 import { AlertService } from './shared/services/alert.service';
-import { HttpClientModule } from '@angular/common/http';
+import { provideHttpClient, withInterceptorsFromDi } from '@angular/common/http';
 import { HomeModule } from './pages/home/home.module';
 
-@NgModule({
-  declarations: [
-    AppComponent
-  ],
-  imports: [
-    BrowserModule,
-    HomeModule,
-    AppRoutingModule,
-    HttpClientModule
-  ],
-  providers:[
-    AlertService
-  ],
-  bootstrap: [AppComponent]
-})
+@NgModule({ declarations: [
+        AppComponent
+    ],
+    bootstrap: [AppComponent], imports: [BrowserModule,
+        HomeModule,
+        AppRoutingModule], providers: [
+        AlertService,
+        provideHttpClient(withInterceptorsFromDi())
+    ] })
 export class AppModule { }
